+++ src/problem/constr.c
@@ -40,9 +40,10 @@ struct Constr {
   Vec* u;           /** @brief Upper bound for linear inequality contraints */
   
   // Extra variables
-  int num_extra_vars; /** @brief Number of extra variables (set during "count") */
-  Vec* l_extra_vars;  /** @brief Lower bounds for extra variables (set during "analyze") */
-  Vec* u_extra_vars;  /** @brief Upper bounds for extra varaibles (set during "analyze") */
+  int num_extra_vars;   /** @brief Number of extra variables (set during "count") */
+  Vec* l_extra_vars;    /** @brief Lower bounds for extra variables (set during "analyze") */
+  Vec* u_extra_vars;    /** @brief Upper bounds for extra varaibles (set during "analyze") */
+  Vec* init_extra_vars; /** @brief Extra variable initial values */
   
   // Counters and flags
   int A_nnz;             /**< @brief Counter for nonzeros of matrix A */
@@ -148,6 +149,7 @@ void CONSTR_del_matvec(Constr* c) {
     VEC_del(c->u);
     VEC_del(c->l_extra_vars);
     VEC_del(c->u_extra_vars);
+    VEC_del(c->init_extra_vars);
     MAT_array_del(c->H_array,c->H_array_size);
     MAT_del(c->H_combined);
     c->b = NULL;
@@ -159,6 +161,7 @@ void CONSTR_del_matvec(Constr* c) {
     c->u = NULL;
     c->l_extra_vars = NULL;
     c->u_extra_vars = NULL;
+    c->init_extra_vars = NULL;
     c->H_array = NULL;
     c->H_array_size = 0;
     c->H_combined = NULL;
@@ -223,6 +226,13 @@ Vec* CONSTR_get_l_extra_vars(Constr* c) {
 Vec* CONSTR_get_u_extra_vars(Constr* c) {
   if (c)
     return c->u_extra_vars;
+  else
+    return NULL;
+}
+
+Vec* CONSTR_get_init_extra_vars(Constr* c) {
+  if (c)
+    return c->init_extra_vars;
   else
     return NULL;
 }
@@ -585,6 +595,7 @@ Constr* CONSTR_new(Net* net) {
   c->u = NULL;
   c->l_extra_vars = NULL;
   c->u_extra_vars = NULL;
+  c->init_extra_vars = NULL;
   c->A_nnz = 0;
   c->J_nnz = 0;
   c->G_nnz = 0;
@@ -651,6 +662,11 @@ void CONSTR_set_l_extra_vars(Constr* c, Vec* l) {
 void CONSTR_set_u_extra_vars(Constr* c, Vec* u) {
   if (c)
     c->u_extra_vars = u;
+}
+
+void CONSTR_set_init_extra_vars(Constr* c, Vec* init) {
+  if (c)
+    c->init_extra_vars = init;
 }
 
 void CONSTR_set_G(Constr* c, Mat* G) {
@@ -855,7 +871,8 @@ BOOL CONSTR_is_safe_to_analyze(Constr* c) {
       MAT_get_size2(CONSTR_get_G(c)) == num_vars &&
       MAT_get_size2(CONSTR_get_J(c)) == num_vars &&
       VEC_get_size(CONSTR_get_l_extra_vars(c)) == CONSTR_get_num_extra_vars(c) &&
-      VEC_get_size(CONSTR_get_u_extra_vars(c)) == CONSTR_get_num_extra_vars(c))
+      VEC_get_size(CONSTR_get_u_extra_vars(c)) == CONSTR_get_num_extra_vars(c) &&
+      VEC_get_size(CONSTR_get_init_extra_vars(c)) == CONSTR_get_num_extra_vars(c))
     return TRUE;
   else {
     sprintf(c->error_string,"constraint is not safe to analyze");