+++ src/problem/constr/constr_AC_FLOW_LIM.c
@@ -253,9 +253,10 @@ void CONSTR_AC_FLOW_LIM_allocate(Constr* c) {
   H_nnz = CONSTR_get_H_nnz(c);
   J_row = CONSTR_get_J_row(c);
 
-  // Extra var limits
+  // Extra vars
   CONSTR_set_l_extra_vars(c,VEC_new(num_extra_vars));
   CONSTR_set_u_extra_vars(c,VEC_new(num_extra_vars));
+  CONSTR_set_init_extra_vars(c,VEC_new(num_extra_vars));
 
   // A b
   CONSTR_set_A(c,MAT_new(0,                       // rows
@@ -332,7 +333,7 @@ void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
   int m;
   int T;
   int num_vars;
-
+  
   // Num periods and net vars
   T = BRANCH_get_num_periods(br);
   num_vars = NET_get_num_vars(CONSTR_get_network(c));
@@ -560,7 +561,7 @@ void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
 
     //**********
 
-    // Extra var limits
+    // Extra vars
     VEC_set(CONSTR_get_l_extra_vars(c),*J_row,0.);
     VEC_set(CONSTR_get_u_extra_vars(c),*J_row,BRANCH_get_ratingA(br));